modified seg projection methods, added sm_prog to check FCN input vector
sm_prog is an independent executable, for the purposes of checking data
on our rosbags instead of UPenn's. it reads a hesai.pcd file, with
different lidar params, and then uses openCV to create a meaningful
range-image for the existing neural network.

appropriate libraries and headers added to cmakelists, various 80char
resizings,

major addition is the tracking of the cv::Mat point indices following
the range-ordering of the Ort input vector, as well as expansion of the
`_hesaiImages` subclass.

TODO:
- [ ] separate ordering logic from projection methods
- [ ] run on rosbag w appropriate topic
- [ ] generate image histograms
- [ ] apply openCV composite intensity filtering
- [x] feed filtered_cloud into pipeline (bad)
- [ ] train new ONNX model? (+ intensity info)